Modularise BetterDescriptions Events
